EHN BIDS filter
and also a lot of refactoring
