+++ giga_connectome/run.py
@@ -94,7 +94,8 @@ def main(argv=None):
         help="A JSON file describing custom BIDS input filters using PyBIDS."
         "We use the same format as described in fMRIPrep documentation: "
         "https://fmriprep.org/en/latest/faq.html#"
-        "how-do-i-select-only-certain-files-to-be-input-to-fmriprep",
+        "how-do-i-select-only-certain-files-to-be-input-to-fmriprep"
+        "However, the query filed should always be 'bold'",
     )
 
     args = parser.parse_args(argv)
